# ui/desktop/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "goose-server",
    "description": "An AI agent",
    "contact": {
      "name": "Block",
      "email": "ai-oss-tools@block.xyz"
    },
    "license": {
      "name": "Apache-2.0"
    },
    "version": "1.28.0"
  },
  "paths": {
...
    "/dictation/models/{model_id}": {
      "delete": {
        "tags": [
          "super::routes::dictation"
        ],
        "operationId": "delete_model",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Model deleted"
          },
          "404": {
            "description": "Model not found or not downloaded"
          },
          "500": {
            "description": "Failed to delete model"
          }
        }
      }
    },
    "/dictation/models/{model_id}/download": {
      "get": {
        "tags": [
          "super::routes::dictation"
        ],
        "operationId": "get_download_progress",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Download progress",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadProgress"
                }
              }
            }
          },
          "404": {
            "description": "Download not found"
          }
        }
      },
      "post": {
        "tags": [
          "super::routes::dictation"
        ],
        "operationId": "download_model",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Download started"
          },
          "400": {
            "description": "Download already in progress"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": [
          "super::routes::dictation"
        ],
        "operationId": "cancel_download",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Download cancelled"
          },
          "404": {
            "description": "Download not found"
          }
        }
      }
    },
    "/dictation/transcribe": {
      "post": {
        "tags": [
          "super::routes::dictation"
        ],
        "operationId": "transcribe_dictation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TranscribeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Audio transcribed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscribeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (bad base64 or unsupported format)"
          },
          "401": {
            "description": "Invalid API key"
          },
          "412": {
            "description": "Provider not configured"
          },
          "413": {
            "description": "Audio file too large (max 50MB)"
          },
          "429": {
            "description": "Rate limit exceeded"
          },
          "500": {
            "description": "Internal server error"
          },
          "502": {
            "description": "Provider API error"
          },
          "503": {
            "description": "Service unavailable"
          },
          "504": {
            "description": "Request timeout"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DictationProvider": {
        "type": "string",
        "enum": [
          "openai",
          "elevenlabs",
          "groq",
          "swiss_ai_platform",
          "local"
        ]
      },
...
      "DownloadProgress": {
        "type": "object",
        "required": [
          "model_id",
          "status",
          "bytes_downloaded",
          "total_bytes",
          "progress_percent"
        ],
        "properties": {
          "bytes_downloaded": {
            "type": "integer",
            "format": "int64",
            "description": "Bytes downloaded so far",
            "minimum": 0
          },
          "error": {
            "type": "string",
            "description": "Error message if failed",
            "nullable": true
          },
          "eta_seconds": {
            "type": "integer",
            "format": "int64",
            "description": "Estimated time remaining in seconds",
            "nullable": true,
            "minimum": 0
          },
          "model_id": {
            "type": "string",
            "description": "Model ID being downloaded"
          },
          "progress_percent": {
            "type": "number",
            "format": "float",
            "description": "Download progress percentage (0-100)"
          },
          "speed_bps": {
            "type": "integer",
            "format": "int64",
            "description": "Download speed in bytes per second",
            "nullable": true,
            "minimum": 0
          },
          "status": {
            "$ref": "#/components/schemas/DownloadStatus"
          },
          "total_bytes": {
            "type": "integer",
            "format": "int64",
            "description": "Total bytes to download",
            "minimum": 0
          }
        }
      },
      "DownloadStatus": {
        "type": "string",
        "enum": [
          "downloading",
          "completed",
          "failed",
          "cancelled"
        ]
      },
...
      "TranscribeRequest": {
        "type": "object",
        "required": [
          "audio",
          "mime_type",
          "provider"
        ],
        "properties": {
          "audio": {
            "type": "string",
            "description": "Base64 encoded audio data"
          },
          "mime_type": {
            "type": "string",
            "description": "MIME type of the audio (e.g., \"audio/webm\", \"audio/wav\")"
          },
          "provider": {
            "$ref": "#/components/schemas/DictationProvider"
          }
        }
      },
      "TranscribeResponse": {
        "type": "object",
        "required": [
          "text"
        ],
        "properties": {
          "text": {
            "type": "string",
            "description": "Transcribed text from the audio"
          }
        }
      },
...
    }
  }
}